+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Changelog
 
 ## Unreleased
 
+Features:
+
+    - [Application] Disable XDebug by default, very much improve performance.
+      Fixes #317
+
 Improvements:
 
     - [Completion] Do not evaluate left operand when completing expression,